allow parent data containers to be accessed from child scopes
